requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
psutil==5.9.8
ping3==4.0.4
speedtest-cli==2.1.3
